Refactor: 폴더 구조를 변경해라
components, containers, redux, pages 분류
