+++ src/containers/CategoriesContainer.jsx
@@ -3,9 +3,9 @@ import { useDispatch, useSelector } from 'react-redux';
 import {
   selectCategory,
   loadRestaurants,
-} from './actions';
+} from '../redux/actions';
 
-import { get } from './utils';
+import { get } from '../utils/utils';
 
 export default function CategoriesContainer() {
   const dispatch = useDispatch();
