+++ src/containers/LoginFormContainer.jsx
@@ -1,15 +1,15 @@
 import { useDispatch, useSelector } from 'react-redux';
 
-import LoginForm from './LoginForm';
-import LogoutForm from './LogoutForm';
+import LoginForm from '../components/LoginForm';
+import LogoutForm from '../components/LogoutForm';
 
 import {
   changeLoginField,
   requestLogin,
   logout,
-} from './actions';
+} from '../redux/actions';
 
-import { get } from './utils';
+import { get } from '../utils/utils';
 
 export default function LoginFormContainer() {
   const dispatch = useDispatch();
