+++ src/containers/RegionsContainer.jsx
@@ -3,9 +3,9 @@ import { useDispatch, useSelector } from 'react-redux';
 import {
   selectRegion,
   loadRestaurants,
-} from './actions';
+} from '../redux/actions';
 
-import { get } from './utils';
+import { get } from '../utils/utils';
 
 export default function RegionsContainer() {
   const dispatch = useDispatch();
